+++ test/fixtures/cookbooks/wintest/recipes/windows_location.rb
@@ -1,15 +1,13 @@
-# Cookbook Name:: vagrant
-# Recipe:: fedora
+# Cookbook Name:: test
+# Test:: windows_location
 
-# Author:: Joshua Timberman <opensource@housepub.org>
-# Copyright:: Copyright (c) 2014, Joshua Timberman
-# License:: Apache License, Version 2.0
+# Copyright 2018 Sous-chefs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,4 +15,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-include_recipe 'vagrant::rhel'
+ENV['VAGRANT_HOME'] = 'C:/Users/altloc'
+node.default['vagrant']['plugins'] = %w(vagrant-ohai vagrant-winrm vagrant-omnibus)
+
+directory 'C:/Users/altloc' do
+  owner 'vagrant'
+  rights :full_control, 'Everyone', applies_to_children: true
+end
+
+include_recipe 'vagrant'